requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.2
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
